搜索页面完成

+++ app/src/main/java/com/ljp/wanandroid/db/SearchDb.kt
@@ -11,23 +11,39 @@ import org.litepal.LitePal
  */
 object SearchDb {
 
+    /**
+     * 历史记录最多显示的条数
+     */
+    private const val maxShowHistory = 20
+
+    /**
+     * 获取历史记录  id 倒序
+     */
     fun getSearchHistory(): MutableList<SearchHotKeyBean> {
-        return LitePal.order("id desc").find(SearchHotKeyBean::class.java) ?: mutableListOf()
+        val list = LitePal.order("id desc").find(SearchHotKeyBean::class.java) ?: mutableListOf()
+        return if (list.size >= maxShowHistory) {
+            list.subList(0, maxShowHistory)
+        } else {
+            list
+        }
     }
 
     fun clearSearchHistory() {
         LitePal.deleteAll(SearchHotKeyBean::class.java)
     }
 
-    fun saveSearchHistory(data: SearchHotKeyBean) {
+    fun saveSearchHistory(data: SearchHotKeyBean): Boolean {
+        val lastData = LitePal.findLast(SearchHotKeyBean::class.java)
+        if (lastData != null && lastData.name == data.name) {
+            return false
+        }
+
         val list = LitePal.where("name = ?", data.name).find(SearchHotKeyBean::class.java)
         list.forEach {
             it.delete()
         }
-        data.save()
+        return data.save()
     }
 
-    fun deleteLastData() {
-        LitePal.findLast(SearchHotKeyBean::class.java).delete()
-    }
+
 }
